moving the push down filters tests to a new file and test implemented and test for cursor with state number
