+++ airbyte-integrations/connectors/source-snowflake/unit_tests/integration/test_table.py
@@ -163,127 +163,6 @@ class FullRefreshTest(TestCase):
         output = _read(_config(), sync_mode=SyncMode.full_refresh)
         assert len(output.records) == 3
 
-
-class FullRefreshPushDownFilterTest(TestCase):
-
-    def setUp(self):
-        self.push_down_filter_name = "higher_than_2"
-        self.where_clause = 'TEST_COLUMN_1>2'
-        self.test_colum_1_index = 1
-        self.push_down_filter = {
-            "name": self.push_down_filter_name,
-            "parent_stream": f"{_SCHEMA}.{_TABLE}",
-            "where_clause": self.where_clause
-        }
-        self._a_record = lambda: a_snowflake_response("response_get_table")
-
-        config_builder = _config()
-        self.config_builder = config_builder.with_push_down_filter(self.push_down_filter)
-
-        sync_mode = SyncMode.full_refresh
-        self.catalog = CatalogBuilder().with_stream(self.push_down_filter_name, sync_mode).build()
-
-    @mock.patch("source_snowflake.streams.snowflake_parent_stream.uuid.uuid4", return_value=_REQUESTID)
-    @mock.patch("source_snowflake.source.SnowflakeJwtAuthenticator")
-    @HttpMocker()
-    def test_given_one_page_when_read_then_return_records(self, uuid_mock, mock_auth, http_mocker: HttpMocker) -> None:
-        _given_get_timezone(http_mocker)
-        _given_read_schema(http_mocker)
-        _given_table_catalog(http_mocker)
-        _given_table_with_primary_keys(http_mocker)
-
-        http_mocker.post(
-            table_request()
-            .with_table(_TABLE)
-            .with_requestID(_REQUESTID)
-            .with_async()
-            .with_where_statement(self.where_clause)
-            .build(),
-            snowflake_response("async_response", FieldPath("statementStatusUrl"))
-            .with_handle(_HANDLE)
-            .build()
-        )
-
-        http_mocker.get(
-            table_request()
-            .with_handle(_HANDLE)
-            .build(is_get=True),
-            snowflake_response("response_get_table")
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=3))
-            .build()
-        )
-
-        output = self._read(self.config_builder, self.catalog)
-        assert len(output.records) == 1
-
-    @mock.patch("source_snowflake.streams.snowflake_parent_stream.uuid.uuid4", return_value=_REQUESTID)
-    @mock.patch("source_snowflake.source.SnowflakeJwtAuthenticator")
-    @HttpMocker()
-    def test_given_three_pages_read_then_return_records(self, uuid_mock, mock_auth, http_mocker: HttpMocker) -> None:
-        _given_get_timezone(http_mocker)
-        _given_read_schema(http_mocker)
-        _given_table_catalog(http_mocker)
-        _given_table_with_primary_keys(http_mocker)
-
-        http_mocker.post(
-            table_request()
-            .with_table(_TABLE)
-            .with_requestID(_REQUESTID)
-            .with_async()
-            .with_where_statement(self.where_clause)
-            .build(),
-            snowflake_response("async_response", FieldPath("statementStatusUrl"))
-            .with_handle(_HANDLE)
-            .build()
-        )
-
-        http_mocker.get(
-            table_request()
-            .with_handle(_HANDLE)
-            .build(is_get=True),
-            snowflake_response("response_get_table")
-            .with_pagination()
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=3))
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=4))
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=5))
-            .build()
-        )
-
-        http_mocker.get(
-            table_request()
-            .with_handle(_HANDLE)
-            .with_partition(1)
-            .build(is_get=True),
-            snowflake_response("response_get_table")
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=6))
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=7))
-            .build()
-        )
-
-        http_mocker.get(
-            table_request()
-            .with_handle(_HANDLE)
-            .with_partition(2)
-            .build(is_get=True),
-            snowflake_response("response_get_table")
-            .with_record(self._a_record().with_field(JsonPath(f"$.[{self.test_colum_1_index}]"), value=8))
-            .build()
-        )
-
-        output = self._read(self.config_builder, self.catalog)
-        assert len(output.records) == 6
-
-    @classmethod
-    def _read(cls,
-              config_builder: ConfigBuilder,
-              catalog: ConfiguredAirbyteCatalog,
-              state: Optional[List[AirbyteStateMessage]] = None,
-              expecting_exception: bool = False
-              ) -> EntrypointOutput:
-        config = config_builder.build()
-        return read(_source(catalog, config, state), config, catalog, state, expecting_exception)
-
-
 class IncrementalTest(TestCase):
     """
     See the documentation about the values of record in the function source_snowflake.schema_builder.format_field
